gestion de permission et de routes

+++ lib/actions/status/index.ts
@@ -1,10 +1,21 @@
-// app/actions/status.ts
 "use server";
 
 import prisma from "@/lib/prisma";
+
 import { StatusFormData } from "@/lib/schema";
 import { Status as PrismaStatus } from "@/app/generated/prisma/client";
 import { Status } from "@/app/(dashboard)/list/status/columns";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { hasPermission } from "@/lib/auth-utils";
+import { checkPermission } from "../auth-helper";
+import { revalidatePath } from "next/cache";
+
+
+
+
+
+
 
 function mapStatus(s: PrismaStatus): Status {
   return {
@@ -19,7 +30,10 @@ function mapStatus(s: PrismaStatus): Status {
    GET
 ========================= */
 export async function getStatus(): Promise<Status[]> {
+  await checkPermission("canViewStatuses");
+  
   const status = await prisma.status.findMany({
+
     orderBy: { createdAt: "desc" },
   });
 
@@ -30,6 +44,8 @@ export async function getStatus(): Promise<Status[]> {
    CREATE
 ========================= */
 export async function createStatusAction(data: StatusFormData) {
+  await checkPermission("canEditStatuses");
+  
   try {
     const status = await prisma.status.create({
       data: {
@@ -38,6 +54,7 @@ export async function createStatusAction(data: StatusFormData) {
       },
     });
 
+    revalidatePath("/");
     return mapStatus(status);
   } catch (err) {
     console.error("Erreur createStatusAction:", err);
@@ -52,6 +69,8 @@ export async function updateStatusAction(
   statusId: string,
   data: Partial<StatusFormData>
 ) {
+  await checkPermission("canEditStatuses");
+  
   try {
     const status = await prisma.status.update({
       where: { id: statusId },
@@ -61,6 +80,7 @@ export async function updateStatusAction(
       },
     });
 
+    revalidatePath("/");
     return mapStatus(status);
   } catch (err) {
     console.error("Erreur updateStatusAction:", err);
@@ -68,16 +88,22 @@ export async function updateStatusAction(
   }
 }
 
+
 /* =========================
    DELETE
 ========================= */
 export async function deleteStatusAction(statusId: string) {
+  await checkPermission("canEditStatuses");
+  
   try {
     const status = await prisma.status.delete({
+
       where: { id: statusId },
     });
 
+    revalidatePath("/");
     return mapStatus(status);
+
   } catch (err) {
     console.error("Erreur deleteStatusAction:", err);
     throw new Error("Impossible de supprimer le statut");
@@ -89,7 +115,10 @@ export const updateOrderStatus = async (
   orderId: string,
   statusId: string | null
 ) => {
+  await checkPermission("canEditOrders");
+  
   try {
+
     const status = statusId
       ? await prisma.status.findUnique({ where: { id: statusId } })
       : null;
@@ -124,10 +153,12 @@ export const updateOrderStatus = async (
       include: { status: true },
     });
 
-   return {
-  ...order,
-  recallAt: order.recallAt ? order.recallAt.toISOString() : null,
-};
+    revalidatePath("/");
+    return {
+      ...order,
+      recallAt: order.recallAt ? order.recallAt.toISOString() : null,
+    };
+
 
   } catch (err) {
     console.error(err);
